Move to test_result.py
